app/requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
configparser==3.5.0
enum34==1.1.6
futures==3.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
protobuf==3.2.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pylint==1.9.3
python-dotenv==0.9.1
singledispatch==3.4.0.3
six==1.11.0
virtualenv==16.1.0
wrapt==1.10.11
